data structure done, i think, writing a function to print it

+++ 2023/d25-attempt/include/part1.h
@@ -1,6 +1,6 @@
 #ifndef PART1_H
 #define PART1_H
 
-unsigned int part1_solution(char *input);
+unsigned long long part1_solution(char *input);
 
 #endif
